# static/swagger.yaml
openapi: 3.0.0
info:
  title: Wardrobe API
  version: 1.0.0
  description: API for managing wardrobe items and outfits
servers:
  - url: http://127.0.0.1:5000
paths:
  /users:
    get:
      summary: Get all users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    username:
                      type: string
    post:
      summary: Create a new user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
  /users/{id}:
    get:
      summary: Get a specific user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A single user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
    put:
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
    delete:
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User deleted

  /users/{user_id}/clothing_items:
    get:
      summary: Get all clothing items for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of clothing items
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    user_id:
                      type: integer
                    type:
                      type: string
                    color:
                      type: string
                    style:
                      type: string
                    season:
                      type: string
    post:
      summary: Create a new clothing item for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                color:
                  type: string
                style:
                  type: string
                season:
                  type: string
      responses:
        '201':
          description: Clothing item created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  type:
                    type: string
                  color:
                    type: string
                  style:
                    type: string
                  season:
                    type: string

  /users/{user_id}/clothing_items/{item_id}:
    get:
      summary: Get a specific clothing item for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A clothing item
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  type:
                    type: string
                  color:
                    type: string
                  style:
                    type: string
                  season:
                    type: string
    put:
      summary: Update a clothing item for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                color:
                  type: string
                style:
                  type: string
                season:
                  type: string
      responses:
        '200':
          description: Clothing item updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  type:
                    type: string
                  color:
                    type: string
                  style:
                    type: string
                  season:
                    type: string
    delete:
      summary: Delete a clothing item for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Clothing item deleted

  /outfits:
    get:
      summary: Get all outfits
      responses:
        '200':
          description: A list of outfits
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    user_id:
                      type: integer
                    items:
                      type: string
    post:
      summary: Create a new outfit
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                items:
                  type: string
      responses:
        '201':
          description: Outfit created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  items:
                    type: string

  /outfits/{id}:
    get:
      summary: Get a specific outfit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: An outfit
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  items:
                    type: string
    put:
      summary: Update an outfit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                items:
                  type: string
      responses:
        '200':
          description: Outfit updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  items:
                    type: string
    delete:
      summary: Delete an outfit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Outfit deleted